requirements.txt
Django>=4.2
djangorestframework>=3.14
mysqlclient>=2.2
django-filter>=23.2
Pillow>=10.0
django-cors-headers>=4.2
gunicorn>=20.1
channels>=4.0
daphne>=4.0
python-decouple>=3.8
whitenoise>=6.0
psycopg2-binary>=2.9

